Style upload button and add hover notification for upload button
Introduced a new "Upload.scss" file for styling related to the uploading functionality. Modified "Upload.tsx" to include a notification that appears when hovering over the "Upload" button, indicating to the user when no file is selected. This provides better user feedback and enhances the user experience. Also installed "framer-motion" for handling smooth animation of the hover notification. Changes to "package.json" and "package-lock.json" are due to this new installation. Cleaned up "page.tsx" by removing an unnecessary extra line for better code readability.