Remove index.css and add Sass dependency
Deleted index.css file as we are transitioning from CSS to Sass for better management and scalability of styles. Sass module has been added for this purpose. Also, replaced the CSS styles with tailwind CSS utility classes in page.tsx for streamlining the styles. Rest of the changes in package-lock.json and package.json are a result of the new Sass installation.