Add drag-and-drop upload feature, improved button UX and new design elements
Introduced a drag-and-drop upload feature in the Upload page to enhance the user experience. Modified Upload.scss to adjust the width and height for responsive layout and added animations for interactive design. Modified Upload.tsx to implement the drag-and-drop function, and refactored UploadButton.tsx to display a notification if no files are selected. New file PopupNotification.tsx created to control the display of this notification. Additionally, changes were made to UploadMethods.tsx to enable file upload from multiple sources such as computer, link, camera, screen, and microphone.