Add upload steps and enhance file selection features
This commit introduces the enumeration of UploadSteps in the Upload component to stratify the process of uploading a file, improving usability. A 'setUploadStep' prop was also added in UploadMethods to interact with these defined steps. To enhance the presentation of file information, a 'convertBytes' function was included to provide a more user-friendly representation of file size. 'Fetch' function in the 'UploadLink' was also simplified to fetch the file directly with the provided URL. These changes overall enhance the separation of concerns making Upload component more comprehensible and maintainable.