Refactor file upload system for better maintainability and user experience
Added several changes to revamp the system for file upload. A new 'addFiles' function was put in place in Upload component to better manage addition of files. Additionally, 'useEffect' hook was employed to dynamically adjust the steps of the upload process depending on the number of files uploaded. Also, a redundant file check was added to avoid upload of files with identical names. Drag-drop feature of file upload was tailored to work alongside these changes in UploadContainer component.